`nix-shell -p` completion
Note: it's currently a bit messy how the "context"/top-level-nix
expression is determined.

Since the both the default and how it is specified are different across
the programs and the -A option is specified as a general option
_nix_attr_paths determines based on the global command_name what the
context should be.

With the -p support it also accepts the top-level as an argument.

I'm not sure what the cleanest way is:
- Keep it like this
- Make -A a program-specific option (always pass top-level as arguemnt)
- Accept an argument which name a function that returns the top-level.
  The common option spec specify the name and each program define this
  function