Fix zsh 5.2 compatibility again
Not sure why I changed this back from
de64c390dbfd25e5acea8fe0c71262f9ec22a9a5, think it was an oversight.

Anyway, macOS now ships with zsh as their default shell, and it can at
least be as old as 5.3 so it's probably a good idea to support it.