Make it possible to complete `nix`-commands on multiple levels
Needed for commands like `nix flake check` where the behavior for `nix
flake <tab>` must be the same as for `nix <tab>`.

This will become relevant in the future as it's planned to namespace
even more-commands[1].

For simplicity I didn't test full recursion, we probably have to fix
this to support e.g. `nix foo bar baz` in the future.

[1] https://github.com/NixOS/nix/pull/3858