Refactor sub-subcommand completion to complete all of their needed options as well
This is a more generic solution which would allow arbitrary nesting of
commands (e.g. `nix foo bar baz`). It is primarily needed now to make
sure that the sub-subcommands get properly interpreted.

This fixes bugs like:

* Missing command-specific options for sub-subcommands like `nix flake
  update --impure`.
* Completion of the arguments for sub-subcommands works now.
* `nix flake --debug <tab>` provides the proper commands.

I tried to document the new behavior with comments to make sure that the
code in here doesn't become even harder to read as it is now.

Pleaes note that I'm not a seasoned ZSH hacker, so I had to use `sed`
and `grep` on a few occasions, improvements are very welcome.